Feature slice with negative offset (#462)
Also adds `LazyFrame::tail` and `LazyFrame::last`

+++ polars/polars-core/src/frame/mod.rs
@@ -1036,7 +1036,7 @@ impl DataFrame {
     }
 
     /// Slice the DataFrame along the rows.
-    pub fn slice(&self, offset: usize, length: usize) -> Result<Self> {
+    pub fn slice(&self, offset: i64, length: usize) -> Result<Self> {
         let col = self
             .columns
             .par_iter()
@@ -1383,7 +1383,7 @@ impl<'a> Iterator for RecordBatchIter<'a> {
         let mut rb_cols = Vec::with_capacity(self.columns.len());
         // take a slice from all columns and add the the current RecordBatch
         self.columns.iter().for_each(|s| {
-            let slice = s.slice(self.idx, length).unwrap();
+            let slice = s.slice(self.idx as i64, length).unwrap();
             rb_cols.push(Arc::clone(&slice.chunks()[0]))
         });
         let rb = RecordBatch::try_new(Arc::clone(&self.schema), rb_cols).unwrap();
